Added example spec for oneOf and allOf
From the openapi documentation there is in particular one example where we can see a combination of oneOf and allOf. 

Everything around polymorphism and inheritance combined for validation AND code generation is quite complex. a not quite intuitive but working way is expressed here https://swagger.io/docs/specification/data-models/oneof-anyof-allof-not/